Split the various components of the digest service into their own files
