+++ lms/services/digest/__init__.py
@@ -1,0 +1,2 @@
+from lms.services.digest.factory import service_factory
+from lms.services.digest.service import DigestService
